+++ src/main/java/com/zappts/CRUDMTG/repository/CardRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 import com.zappts.CRUDMTG.model.Card;
 
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
-	
+//	List<Card> findByIdcard(Integer idcard);
 	List<Card> findByNamecard(@Param(value = "namecard") String namecard);
 	List<Card> findByedition(@Param(value = "edition") String edition);
 	List<Card> findByFoil(@Param(value = "foil") Boolean foil);
