+++ src/main/java/com/zappts/CRUDMTG/repository/ListsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.zappts.CRUDMTG.model.Lists;
 
-public interface ListsRepository extends JpaRepository<Lists, Long> {
+public interface ListsRepository extends JpaRepository<Lists, Integer> {
 
 
 }
