+++ src/main/java/com/zappts/CRUDMTG/repository/PlayerRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.zappts.CRUDMTG.model.Player;
 
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
 
-	List<Player> findByIdplayer(Long idplayer);
+	List<Player> findByIdplayer(Integer idplayer);
 	
 	
 	// List<Player> findByPlayernamePlayer(String nameplayer);
